jbig2dec: Cast to size_t to avoid overflow.

+++ jbig2_image.c
@@ -460,7 +460,7 @@ jbig2_image_clear(Jbig2Ctx *ctx, Jbig2Image *image, int value)
 {
     const uint8_t fill = value ? 0xFF : 0x00;
 
-    memset(image->data, fill, image->stride * image->height);
+    memset(image->data, fill, (size_t) image->stride * image->height);
 }
 
 /* look up a pixel value in an image.
